Addressed comments. Moved logic from separate file into the propagate_empty.rs
