+++ ballista/scheduler/src/state/aqe/optimizer_rule/mod.rs
@@ -18,7 +18,6 @@
 pub mod coalesce_partitions;
 pub mod datafusion_patch;
 pub mod distributed_exchange;
-pub mod join_info;
 pub mod propagate_empty;
 
 pub use coalesce_partitions::*;
