+++ ballista/scheduler/src/state/aqe/optimizer_rule/propagate_empty.rs
@@ -16,7 +16,6 @@
 // under the License.
 
 use crate::state::aqe::execution_plan::ExchangeExec;
-use crate::state::aqe::optimizer_rule::join_info::{as_join, is_guaranteed_empty};
 use datafusion::common::JoinType;
 use datafusion::common::stats::Precision;
 use datafusion::common::tree_node::{Transformed, TreeNode};
@@ -28,6 +27,7 @@ use datafusion::physical_plan::aggregates::AggregateExec;
 use datafusion::physical_plan::coalesce_batches::CoalesceBatchesExec;
 use datafusion::physical_plan::empty::EmptyExec;
 use datafusion::physical_plan::filter::FilterExec;
+use datafusion::physical_plan::joins::{HashJoinExec, SortMergeJoinExec};
 use datafusion::physical_plan::limit::{GlobalLimitExec, LocalLimitExec};
 use datafusion::physical_plan::projection::ProjectionExec;
 use std::sync::Arc;
@@ -40,8 +40,18 @@ macro_rules! is_empty_exec {
 
 macro_rules! empty_exec {
     ($e:expr) => {
-        Ok(Transformed::yes(Arc::new(EmptyExec::new($e.schema()))))
+        Ok(Transformed::yes(Arc::new(
+            EmptyExec::new($e.schema())
+                .with_partitions($e.properties().output_partitioning().partition_count()),
+        )))
     };
+}
+
+/// Holding information about the subsequent join operation
+pub struct JoinInfo<'a> {
+    pub join_type: JoinType,
+    pub left: &'a Arc<dyn ExecutionPlan>,
+    pub right: &'a Arc<dyn ExecutionPlan>,
 }
 
 /// This is [datafusion::optimizer::propagate_empty_relation::PropagateEmptyRelation] rule with difference
@@ -89,6 +99,7 @@ impl PropagateEmptyExecRule {
             let left_empty = is_guaranteed_empty(join.left);
             let right_empty = is_guaranteed_empty(join.right);
 
+            // Checking whether join would produce an empty result
             let should_eliminate = match join.join_type {
                 JoinType::Inner => left_empty || right_empty,
                 JoinType::Full => left_empty && right_empty,
@@ -103,7 +114,11 @@ impl PropagateEmptyExecRule {
             };
 
             if should_eliminate {
-                Ok(Transformed::yes(Arc::new(EmptyExec::new(plan.schema()))))
+                Ok(Transformed::yes(Arc::new(
+                    EmptyExec::new(plan.schema()).with_partitions(
+                        plan.properties().output_partitioning().partition_count(),
+                    ),
+                )))
             } else {
                 Ok(Transformed::no(plan))
             }
@@ -138,12 +153,194 @@ impl PhysicalOptimizerRule for PropagateEmptyExecRule {
     }
 }
 
+/// Extracting the physical join information from the incoming [ExecutionPlan]
+pub fn as_join(plan: &Arc<dyn ExecutionPlan>) -> Option<JoinInfo<'_>> {
+    let any = plan.as_any();
+
+    if let Some(join) = any.downcast_ref::<HashJoinExec>() {
+        return Some(JoinInfo {
+            join_type: join.join_type,
+            left: join.left(),
+            right: join.right(),
+        });
+    }
+    if let Some(join) = any.downcast_ref::<SortMergeJoinExec>() {
+        return Some(JoinInfo {
+            join_type: join.join_type,
+            left: join.left(),
+            right: join.right(),
+        });
+    }
+
+    None
+}
+
+/// Getting sure that incoming plan (either from left or right) is empty
+///
+/// Note that when we have [Precision::Absent] it means we know nothing about the value
+/// Hence we can't decide whether it is empty
+pub fn is_guaranteed_empty(plan: &Arc<dyn ExecutionPlan>) -> bool {
+    let Ok(stats) = plan.partition_statistics(None) else {
+        return false;
+    };
+
+    match stats.num_rows {
+        Precision::Exact(n) => n == 0,
+        _ => false,
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::state::aqe::optimizer_rule::join_info::test_helpers::*;
-    use datafusion::common::JoinType;
 
+    use datafusion::{
+        arrow::datatypes::{DataType, Field, Schema},
+        common::{ColumnStatistics, JoinType, NullEquality, Statistics},
+        physical_plan::{
+            expressions::Column, joins::PartitionMode, test::exec::StatisticsExec,
+        },
+    };
+
+    /// Helpers
+    pub fn schema() -> Arc<Schema> {
+        Arc::new(Schema::new(vec![Field::new("a", DataType::Int32, true)]))
+    }
+
+    pub fn empty_stats_exec() -> Arc<dyn ExecutionPlan> {
+        Arc::new(StatisticsExec::new(
+            Statistics {
+                num_rows: Precision::Exact(0),
+                total_byte_size: Precision::Exact(0),
+                column_statistics: vec![ColumnStatistics::new_unknown()],
+            },
+            Schema::new(vec![Field::new("a", DataType::Int32, true)]),
+        ))
+    }
+
+    pub fn non_empty_stats_exec() -> Arc<dyn ExecutionPlan> {
+        Arc::new(StatisticsExec::new(
+            Statistics {
+                num_rows: Precision::Exact(100),
+                total_byte_size: Precision::Exact(800),
+                column_statistics: vec![ColumnStatistics::new_unknown()],
+            },
+            Schema::new(vec![Field::new("a", DataType::Int32, true)]),
+        ))
+    }
+
+    pub fn unknown_stats_exec() -> Arc<dyn ExecutionPlan> {
+        Arc::new(StatisticsExec::new(
+            Statistics {
+                num_rows: Precision::Absent,
+                total_byte_size: Precision::Absent,
+                column_statistics: vec![ColumnStatistics::new_unknown()],
+            },
+            Schema::new(vec![Field::new("a", DataType::Int32, true)]),
+        ))
+    }
+
+    pub fn join_on() -> Vec<(
+        Arc<dyn datafusion::physical_plan::PhysicalExpr>,
+        Arc<dyn datafusion::physical_plan::PhysicalExpr>,
+    )> {
+        vec![(Arc::new(Column::new("a", 0)), Arc::new(Column::new("a", 0)))]
+    }
+
+    pub fn hash_join(
+        left: Arc<dyn ExecutionPlan>,
+        right: Arc<dyn ExecutionPlan>,
+        join_type: JoinType,
+    ) -> Arc<dyn ExecutionPlan> {
+        Arc::new(
+            HashJoinExec::try_new(
+                left,
+                right,
+                join_on(),
+                None,
+                &join_type,
+                None,
+                PartitionMode::Partitioned,
+                NullEquality::NullEqualsNothing,
+                false,
+            )
+            .unwrap(),
+        )
+    }
+
+    pub fn sort_merge_join(
+        left: Arc<dyn ExecutionPlan>,
+        right: Arc<dyn ExecutionPlan>,
+        join_type: JoinType,
+    ) -> Arc<dyn ExecutionPlan> {
+        let sort_options = vec![Default::default()];
+        Arc::new(
+            SortMergeJoinExec::try_new(
+                left,
+                right,
+                join_on(),
+                None,
+                join_type,
+                sort_options,
+                NullEquality::NullEqualsNothing,
+            )
+            .unwrap(),
+        )
+    }
+
+    #[test]
+    fn guaranteed_empty_exact_zero() {
+        assert!(is_guaranteed_empty(&empty_stats_exec()));
+    }
+
+    #[test]
+    fn guaranteed_empty_non_zero() {
+        assert!(!is_guaranteed_empty(&non_empty_stats_exec()));
+    }
+
+    #[test]
+    fn guaranteed_empty_unknown_stats() {
+        // Precision::Absent — must NOT claim empty, otherwise we would make false claims about it being empty
+        assert!(!is_guaranteed_empty(&unknown_stats_exec()));
+    }
+
+    #[test]
+    fn guaranteed_empty_literal_empty_exec() {
+        let plan = Arc::new(EmptyExec::new(schema())) as Arc<dyn ExecutionPlan>;
+        assert!(is_guaranteed_empty(&plan));
+    }
+
+    #[test]
+    fn as_join_recognises_hash_join() {
+        let plan = hash_join(
+            non_empty_stats_exec(),
+            non_empty_stats_exec(),
+            JoinType::Inner,
+        );
+        let info = as_join(&plan);
+        assert!(info.is_some());
+        assert_eq!(info.unwrap().join_type, JoinType::Inner);
+    }
+
+    #[test]
+    fn as_join_recognises_sort_merge_join() {
+        let plan = sort_merge_join(
+            non_empty_stats_exec(),
+            non_empty_stats_exec(),
+            JoinType::Left,
+        );
+        let info = as_join(&plan);
+        assert!(info.is_some());
+        assert_eq!(info.unwrap().join_type, JoinType::Left);
+    }
+
+    #[test]
+    fn as_join_returns_none_for_non_join() {
+        let plan = non_empty_stats_exec();
+        assert!(as_join(&plan).is_none());
+    }
+
+    // Testing correctness of empty join eliminating logic
     #[test]
     fn inner_left_empty_eliminated() {
         let plan = hash_join(empty_stats_exec(), non_empty_stats_exec(), JoinType::Inner);
